+++ 2017-10/ClientAPIs/adam/__init__.py
@@ -8,6 +8,7 @@ from adam.batch import Batches
 from adam.batch_runner import BatchRunner
 from adam.group import Groups
 from adam.project import Projects
+from adam.service import Service
 from adam.timer import Timer
 from adam.rest_proxy import RestRequests
 from adam.rest_proxy import AuthorizingRestProxy
